Draft of featurizer refactor, with audio to [spectrogram, mel spect, mfcc] modules.
