+++ collections/nemo_asr/nemo_asr/parts/features.py
@@ -3,7 +3,6 @@
 import math
 import librosa
 import torch
-import torchaudio
 import torch.nn as nn
 from .perturb import AudioAugmentor
 from .segment import AudioSegment
@@ -97,142 +96,9 @@ class FeaturizerFactory(object):
             perturbation_configs=perturbation_configs)
 
 
-class SpectrogramFeatures(nn.Module):
-    def __init__(self, sample_rate=8000, window_size=0.02, window_stride=0.01,
-                 n_fft=None,
-                 window="hamming", normalize=True, log=True, center=True,
-                 dither=CONSTANT, pad_to=8, max_duration=16.7,
-                 frame_splicing=1):
-        super(SpectrogramFeatures, self).__init__()
-        torch_windows = {
-            'hann': torch.hann_window,
-            'hamming': torch.hamming_window,
-            'blackman': torch.blackman_window,
-            'bartlett': torch.bartlett_window,
-            'none': None,
-        }
-        self.win_length = int(sample_rate * window_size)
-        self.hop_length = int(sample_rate * window_stride)
-        self.n_fft = n_fft or 2 ** math.ceil(math.log2(self.win_length))
-
-        window_fn = torch_windows.get(window, None)
-        window_tensor = window_fn(self.win_length,
-                                  periodic=False) if window_fn else None
-        self.window = window_tensor
-
-        self.normalize = normalize
-        self.log = log
-        self.center = center
-        self.dither = dither
-        self.pad_to = pad_to
-        self.frame_splicing = frame_splicing
-
-    def get_seq_len(self, seq_len):
-        return torch.ceil(seq_len.to(dtype=torch.float) / self.hop_length).to(
-            dtype=torch.long)
-
-    @torch.no_grad()
-    def forward(self, x, seq_len):
-        dtype = x.dtype
-        x = x.to(torch.float)
-
-        seq_len = self.get_seq_len(seq_len)
-
-        # dither
-        if self.dither > 0:
-            x += self.dither * torch.randn_like(x)
-
-        # do preemphasis
-        if hasattr(self, 'preemph') and self.preemph is not None:
-            x = torch.cat(
-                (x[:, 0].unsqueeze(1), x[:, 1:] - self.preemph * x[:, :-1]),
-                dim=1)
-
-        # get spectrogram
-        x = torch.stft(x, n_fft=self.n_fft, hop_length=self.hop_length,
-                       win_length=self.win_length, center=self.center,
-                       window=self.window.to(torch.float))
-        x = torch.sqrt(x.pow(2).sum(-1))
-
-        # log features if required
-        if self.log:
-            x = torch.log(x + 1e-20)
-
-        # frame splicing if required
-        if self.frame_splicing > 1:
-            x = splice_frames(x, self.frame_splicing)
-
-        # normalize if required
-        if self.normalize:
-            x = normalize_batch(x, seq_len, normalize_type=self.normalize)
-
-        # mask to zero any values beyond seq_len in batch, pad to multiple of
-        # `pad_to` (for efficiency)
-        max_len = x.size(-1)
-        mask = torch.arange(max_len).to(seq_len.device)
-        mask = mask.expand(x.size(0), max_len) >= seq_len.unsqueeze(1)
-        x = x.masked_fill(mask.unsqueeze(1).to(device=x.device), 0)
-        del mask
-        if self.pad_to > 0:
-            pad_amt = x.size(-1) % self.pad_to
-            if pad_amt != 0:
-                x = nn.functional.pad(x, (0, self.pad_to - pad_amt))
-
-        return x.to(dtype)
-
-    @classmethod
-    def from_config(cls, cfg, log=False):
-        return cls(sample_rate=cfg['sample_rate'],
-                   window_size=cfg['window_size'],
-                   window_stride=cfg['window_stride'],
-                   n_fft=cfg['n_fft'], window=cfg['window'],
-                   normalize=cfg['normalize'],
-                   dither=cfg.get('dither', 1e-5), pad_to=cfg.get("pad_to", 0),
-                   frame_splicing=cfg.get("frame_splicing", 1), log=log)
-
-
 class FilterbankFeatures(nn.Module):
     """Featurizer that converts wavs to Mel Spectrograms.
-
-    Args:
-        sample_rate (int): Sample rate of the input audio data.
-            Defaults to 16000
-        window_size (float): Size of window for fft in seconds
-            Defaults to 0.02
-        window_stride (float): Stride of window for fft in seconds
-            Defaults to 0.01
-        window (str): Windowing function for fft. can be one of ['hann',
-            'hamming', 'blackman', 'bartlett']
-            Defaults to "hann"
-        normalize (str): Can be one of ['per_feature', 'all_features']; all
-            other options disable feature normalization. 'all_features'
-            normalizes the entire spectrogram to be mean 0 with std 1.
-            'pre_features' normalizes per channel / freq instead.
-            Defaults to "per_feature"
-        n_fft (int): Length of FT window. If None, it uses the smallest power
-            of 2 that is larger than n_window_size.
-            Defaults to None
-        preemph (float): Amount of pre emphasis to add to audio. Can be
-            disabled by passing None.
-            Defaults to 0.97
-        nfilt (int): Number of mel spectrogram freq bins to output.
-            Defaults to 64
-        lowfreq (int): Lower bound on mel basis in Hz.
-            Defaults to 0
-        highfreq  (int): Lower bound on mel basis in Hz.
-            Defaults to None
-        log (bool): Log features.
-            Defaults to True
-        dither (float): Amount of white-noise dithering.
-            Defaults to 1e-5
-        pad_to (int): Ensures that the output size of the time dimension is
-            a multiple of pad_to.
-            Defaults to 16
-        max_duration (float): Loose ceiling on max audio duration
-        frame_splicing (int): Defaults to 1
-        stft_conv (bool): If True, uses pytorch_stft and convolutions. If
-            False, uses torch.stft.
-            Defaults to False
+    See AudioToMelSpectrogramPreprocessor for args.
     """
     def __init__(
             self, *,
@@ -401,79 +267,3 @@ class FilterbankFeatures(nn.Module):
                    frame_splicing=cfg.get("frame_splicing", 1),
                    log=log,
                    stft_conv=cfg.get("stft_conv", False))
-
-
-class MFCCFeatures(nn.Module):
-    """Featurizer that converts wavs to MFCCs.
-    Uses torchaudio.transforms.MFCC, and arguments are the same.
-
-    Args:
-        sample_rate: The sample rate of the audio
-        n_mfcc: Number of coefficients to retain
-        dct_type: Type of discrete cosine transform to use
-        norm: Type of norm to use
-        log: Whether to use log-mel spectrograms instead of db-scaled.
-            Defaults to True.
-        mel_kwargs: Dict of arguments for torchaudio.transforms.MelSpectrogram
-    """
-    def __init__(
-            self, *,
-            sample_rate=16000,
-            n_mfcc=40,
-            dct_type=2,
-            norm='ortho',
-            log=True,
-            mel_kwargs=None
-    ):
-        super(MFCCFeatures, self).__init__()
-
-        # Use the sample rate given to MFCCFeatures instead of in mel_kwargs
-        if mel_kwargs and 'sample_rate' in mel_kwargs:
-            del mel_kwargs['sample_rate']
-
-        self.mfcc = torchaudio.transforms.MFCC(
-            sample_rate=sample_rate,
-            n_mfcc=n_mfcc,
-            dct_type=dct_type,
-            norm=norm,
-            log_mels=log,
-            melkwargs=mel_kwargs
-        )
-
-    def get_seq_len(self, seq_len):
-        return torch.ceil(
-                seq_len.to(dtype=torch.float) /
-                self.mfcc.MelSpectrogram.hop_length).to(dtype=torch.long)
-
-    @torch.no_grad()
-    def forward(self, x, seq_len):
-        return self.mfcc(x)
-
-    @classmethod
-    def from_config(cls, cfg, log=False):
-        return cls(sample_rate=cfg.get('sample_rate', 16000),
-                   n_mfcc=cfg.get('n_mfcc', 40),
-                   dct_type=cfg.get('dct_type', 2),
-                   norm=cfg.get('norm', 'ortho'),
-                   log=log,
-                   mel_kwargs=cfg.get('mel_kwargs', None))
-
-
-class FeatureFactory(object):
-    featurizers = {
-        "logfbank": FilterbankFeatures,
-        "fbank": FilterbankFeatures,
-        "stft": SpectrogramFeatures,
-        "logspect": SpectrogramFeatures,
-        "logstft": SpectrogramFeatures,
-        "mfcc": MFCCFeatures
-    }
-
-    def __init__(self):
-        pass
-
-    @classmethod
-    def from_config(cls, cfg):
-        feat_type = cfg.get('feat_type', "logspect")
-        featurizer = cls.featurizers[feat_type]
-        return featurizer.from_config(cfg, log="log" in feat_type)
